+++ src/lib.rs
@@ -35,6 +35,7 @@ pub mod session;
 pub mod shell;
 pub mod signal;
 pub mod signature;
+pub mod ssh_entry;
 pub mod stream;
 pub mod structured;
 pub mod suggest;